project/backend/requirements.txt
# 使用清华源安装: pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
# 训练相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
torch>=2.2.0
torchvision>=0.17.0
ultralytics==8.0.196
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
# 训练和评估相关
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1 